+++ ricaun.RevitTest.Application/Module.cs
@@ -8,7 +8,9 @@ namespace ricaun.RevitTest.Application
 {
     class Module
     {
+#pragma warning disable CA2255 // The 'ModuleInitializer' attribute should not be used in libraries
         [ModuleInitializer]
+#pragma warning restore CA2255 // The 'ModuleInitializer' attribute should not be used in libraries
         internal static void Initialize()
         {
             //Debug.WriteLine($"Module: {typeof(Module).Assembly}");
@@ -21,8 +23,10 @@ namespace ricaun.RevitTest.Application
     }
 }
 
+#if NETFRAMEWORK
 namespace System.Runtime.CompilerServices
 {
     [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
     internal sealed class ModuleInitializerAttribute : Attribute { }
 }
+#endif