+++ ricaun.RevitTest.Application/Revit/ApsApplication/ApsApplication.cs
@@ -1,6 +1,7 @@
 ﻿using ricaun.Aps;
 using ricaun.Aps.Extensions;
 using ricaun.Auth.Aps;
+using System.Diagnostics;
 using System.Threading.Tasks;
 
 namespace ricaun.RevitTest.Application.Revit.ApsApplication
@@ -16,17 +17,58 @@ namespace ricaun.RevitTest.Application.Revit.ApsApplication
         public static ApsService ApsService;
         public static bool IsConnected => ApsService?.IsConnected ?? false;
         public static string LoginUserId => GetLoginUserId();
-        public static string GetLoginUserId()
+        private static string GetLoginUserId()
         {
             if (IsConnected == false)
                 return string.Empty;
 
-            var data = ApsService.ApsClient?.GetAccessToken()?.GetDataBase();
-
+            var data = ApsService.ApsClient?.GetOpenIdData();
             if (data is null)
-                return string.Empty;
+            {
+                Debug.WriteLine($"OpenIdData: is null.");
+                var tokenData = ApsService.ApsClient?.GetAccessToken()?.GetAccessTokenData();
+                if (tokenData is null)
+                {
+                    return string.Empty;
+                }
+                return tokenData.UserId;
+            }
 
             return data.UserId;
+        }
+
+        public static async Task<bool> EnsureApsUserHaveOpenId()
+        {
+            if (IsConnected == false)
+                return false;
+
+            //async Task DisconnectUser()
+            //{
+            //    Debug.WriteLine($"EnsureApsUserHaveOpenId: Force to disconnect User.");
+            //    await Logout();
+            //}
+
+            var tokenData = ApsService.ApsClient?.GetAccessToken()?.GetAccessTokenData();
+            if (tokenData is null)
+            {
+                Debug.WriteLine($"EnsureApsUserHaveOpenId: AccessTokenData is empty.");
+                //await DisconnectUser();
+                //return false;
+            }
+
+            Debug.WriteLine($"AccessTokenData Timeout: {tokenData.Timeout()}.");
+
+            var data = ApsService.ApsClient?.GetOpenIdData();
+            if (data is null)
+            {
+                Debug.WriteLine($"EnsureApsUserHaveOpenId: OpenIdData is empty.");
+                //await DisconnectUser();
+                //return false;
+            }
+
+            await Task.Delay(0);
+
+            return true;
         }
         public static async Task Initialize()
         {