+++ ricaun.RevitTest.Application/Revit/ApsApplication/Extensions/TaskExtension.cs
@@ -1,0 +1,16 @@
+﻿using System.Threading.Tasks;
+
+namespace ricaun.RevitTest.Application.Revit.ApsApplication
+{
+    public static class TaskExtension
+    {
+        public static async void RunAndForget(this Task task)
+        {
+            try
+            {
+                await task;
+            }
+            catch { }
+        }
+    }
+}