+++ ricaun.RevitTest.Application/Revit/Log.cs
@@ -78,7 +78,15 @@ namespace ricaun.RevitTest.Application.Revit
         {
             if (File.Exists(FilePath))
             {
-                Process.Start(FilePath);
+                try
+                {
+                    Process.Start(FilePath);
+                }
+                catch (Exception)
+                {
+                    Process.Start(new ProcessStartInfo(FilePath) { UseShellExecute = true });
+                    //Process.Start(new ProcessStartInfo("cmd", $"/c start {FilePath}"));
+                }
             }
         }
         #endregion