+++ ricaun.RevitTest.Console/Revit/Utils/VisualStudioDebugUtils.cs
@@ -49,7 +49,7 @@ namespace ricaun.RevitTest.Console.Revit.Utils
                 try
                 {
                     var objectDTE = $"VisualStudio.DTE.{i}.0";
-                    var dte = (DTE)System.Runtime.InteropServices.Marshal.GetActiveObject(objectDTE);
+                    var dte = GetActiveDTEObject(objectDTE);
                     if (dte is not null)
                     {
                         Debug.WriteLine($"DTE.Debugger: {objectDTE}");
@@ -59,6 +59,15 @@ namespace ricaun.RevitTest.Console.Revit.Utils
                 catch { }
             }
             return null;
+        }
+
+        private static DTE GetActiveDTEObject(string objectDTE)
+        {
+#if NETFRAMEWORK
+            return (DTE)System.Runtime.InteropServices.Marshal.GetActiveObject(objectDTE);
+#elif NETCOREAPP
+            return (DTE)MarshalUtils.GetActiveObject(objectDTE);
+#endif
         }
 
         public static string GetName()
@@ -72,5 +81,4 @@ namespace ricaun.RevitTest.Console.Revit.Utils
             return string.Format("{0} {1}", dte.Name, dte.Version);
         }
     }
-
 }